refactor : EditProduct와 NewProduct를 합친 SendingProduct 모델 생성
- SendingProduct로 합치면서 생긴 사이드 이팩트 해결
